// src/interfaces/keycodes.ts
export enum keycodes {
    Left = 37,
    Up = 38,
    Right = 39,
    Down = 40,
    Enter = 13,
    Backspace = 8,
    Delete = 46,
    Two = 50,
    Escape = 27,
    Dash = 189,
    Eight = 56,
    Tab = 9,
    Shift = 16,
    Semicolon = 186,
    Space = 32,
    B = 66,
    I = 73,
    U = 85,
    Z = 90,
    Y = 89,
    S = 83,
    Period = 190,
    Zero = 48,
    Equal = 187,
    Minus = 189
}